pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.175.0",
    "google-search-results>=2.4.2",
    "litellm>=1.74.9",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "tavily-python>=0.7.11",
]
